% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/illustrative-analysis-helpers.R
\name{choose_standard_refpat}
\alias{choose_standard_refpat}
\title{Choose standard reference patient to predict}
\usage{
choose_standard_refpat(
  col,
  contin_action = c("median", "mean"),
  categ_action = c("most_common", "reference")
)
}
\arguments{
\item{col}{A vector, generally corresponding to a dataframe column}

\item{contin_action}{Character, choosing whether to select mean or median value
for a continuous variable}

\item{categ_action}{Character, choosing whether to select most common or reference
factor levels for a categorical variable}
}
\value{
Returns single either numeric or character value for a specific col
}
\description{
Based on either most common or reference factor levels in sample for
categorical variable, or median/mean for continuous variables.
}
