% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/illustrative-analysis-helpers.R
\name{cloglog}
\alias{cloglog}
\title{Complementary log-log transformation}
\usage{
cloglog(x)
}
\arguments{
\item{x}{Scalar or numeric vector}
}
\description{
Complementary log-log transformation
}
