% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra_helpers.R
\name{cumincs_plot_truepred}
\alias{cumincs_plot_truepred}
\title{Visualise true/predicted probs with covariates}
\usage{
cumincs_plot_truepred(cox_long, combo, ev1_pars, ev2_pars, dat)
}
\arguments{
\item{cox_long}{Cox model returned by setup_mstate}

\item{combo}{Covariate combo, e.g. list("val_X" = 1, "val_Z" = 1)}

\item{ev1_pars}{List arameters for weibull event 1}

\item{ev2_pars}{List arameters for weibull event 2}

\item{dat}{Dataset as generated by generate_dat()}
}
\value{
Ggplot by state and true/predicted
}
\description{
Visualise true/predicted probs with covariates
}
