% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{dat_mds_synth}
\alias{dat_mds_synth}
\title{Synthetic version of MDS long-term data}
\format{
\describe{
\item{ci_s_allo1}{Competing event indicator: 0 = censored, 1 = relapse,
2 = non-relapse mortality.}
\item{ci_allo1}{Time to first competing event (years).}
\item{match_allo1_1}{Sex match patient and donor. Unordered categorical with four levels,
corresponding to all combinations of recipient male/female and donor male/female.}
\item{mdsclass}{MDS groups based on subclassification at alloHCT. Categorical with three
levels: MDS without excess blasts, with excess blast and sAML.}
\item{donorrel}{Human leukocyte antigen match between patient and donor. Binary: 
Identical sibling or Other.}
\item{karnofsk_allo1}{Karnofsky performance status. Three-level ordered categorical:
>=90, 80 and <=70.}
\item{crnocr}{Stage at alloHCT. Three-level unordered categorical: CR, 
 noCR and Untreated/not aimed at remission.}
\item{cmv_combi_allo1_1}{Cytomegalovirus (CMV) status in patient and donor. 
Four-level categorical corresponding to all combinations of patient positive/negative
and donor positive/negative. For example, level "+/-" corresponds to CMV positive
in patient and negative in donor.}
\item{cytog_threecat}{Cytogenetics categories within International Prognostic 
Scoring System}
\item{hctci_risk}{Hematopoietic stemcell transplantation-comorbidity index score.
Three-level ordered categorical: low risk (0), intermediate risk (1-2) and 
high risk (>= 3).}
\item{agedonor_allo1_decades}{Donor age at alloHCT (decades).}
\item{age_allo1_decades}{Patient age at alloHCT (decades).}
}
}
\usage{
dat_mds_synth
}
\description{
Synthetic version of MDS long-term data
}
\keyword{datasets}
