% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_generation_helpers.R
\name{gen_cmprsk_times}
\alias{gen_cmprsk_times}
\title{Generate competing risks times + event indictor (2 events)}
\usage{
gen_cmprsk_times(n, .data, ev1_pars, ev2_pars, rate_cens)
}
\arguments{
\item{n}{Sample size.}

\item{.data}{Data frame containing n rows with X and Z as columns}

\item{ev1_pars}{Named list of ("a1", "h1_0", "b1", "gamm1")}

\item{ev2_pars}{Named list of ("a2", "h2_0", "b2", "gamm2")}

\item{rate_cens}{Rate of exponential distributiion for censoring.
If 0 means no censoring is applied.}
}
\value{
Data-frame with missings induced
}
\description{
Generate competing risks times + event indictor (2 events)
}
