% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-generation.R
\name{generate_dat}
\alias{generate_dat}
\title{Generates data.}
\usage{
generate_dat(
  n,
  X_type,
  r,
  ev1_pars,
  ev2_pars,
  rate_cens,
  mech = NULL,
  eta1 = NULL,
  p = NULL,
  mod_type = "latent"
)
}
\arguments{
\item{n}{Sample size.}

\item{X_type}{Either "binary", or "contin"}

\item{r}{Desired correlation between X and Z.}

\item{ev1_pars}{Named list of ("a1" =, "h1_0", "b1", "gamm1")}

\item{ev2_pars}{Named list of ("a2", "h2_0", "b2", "gamm2")}

\item{rate_cens}{Rate of exponential distributiion for censoring.
If 0 means no censoring is applied.}

\item{mech}{Missingness mechanism, one of:
"MAR_GEN", "MAR", "MNAR", "MCAR"}

\item{eta1}{Only necessary for mech != "MCAR" - degree/direction
of assocation between variable responsible for missingness
and probability of missingness}

\item{p}{Proportion of missing values in X.}

\item{mod_type}{Either "latent", weibull times generated from 
separate weibull distribution, or "total", times generated from 
sum of cause-specific hazards (using inverse transform method). For educational
purposes - both methods yield virtually same results.}
}
\value{
Data-frame with missings induced
}
\description{
Generates data.
}
