% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MI_helpers.R
\name{get_imp_models}
\alias{get_imp_models}
\title{Get vector of imputation models to use based on class of X}
\usage{
get_imp_models(dat)
}
\arguments{
\item{dat}{Dataset as generated by generate_dat()}
}
\value{
Named character vector with methods to impute
}
\description{
Get vector of imputation models to use based on class of X
}
