% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MI_helpers.R
\name{get_predictor_mats}
\alias{get_predictor_mats}
\title{Prepare prediction matrices MICE/smcfcs}
\usage{
get_predictor_mats(dat)
}
\arguments{
\item{dat}{Dataset as generated by generate_dat()}
}
\value{
List of prediction matrices
}
\description{
Prepare prediction matrices MICE/smcfcs
}
