% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_helpers.R
\name{get_preds_grid}
\alias{get_preds_grid}
\title{Predicting grids}
\usage{
get_preds_grid(cox_long, grid_obj, times, ev1_pars, ev2_pars)
}
\arguments{
\item{cox_long}{Cox model returned by setup_mstate}

\item{grid_obj}{Grid object}

\item{times}{Prediction horizon(s) eg. c(2, 5, 10)}

\item{ev1_pars}{List arameters for weibull event 1}

\item{ev2_pars}{List arameters for weibull event 2}
}
\value{
Df of predictions
}
\description{
Predicting grids
}
