% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstate_helpers.R
\name{get_true_cuminc}
\alias{get_true_cuminc}
\title{Compute true cumulative incidence}
\usage{
get_true_cuminc(ev1_pars, ev2_pars, combo, times)
}
\arguments{
\item{ev1_pars}{List arameters for weibull event 1}

\item{ev2_pars}{List arameters for weibull event 2}

\item{combo}{Covariate combo, e.g. list("val_X" = 1, "val_Z" = 1)}

\item{times}{Vector of timepoints to evaluate cuminc at}
}
\value{
True cumulative incidence at times.
}
\description{
Compute true cumulative incidence
}
