% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{ggplot_nlp}
\alias{ggplot_nlp}
\title{Custom ggplot-based nested loop plots}
\usage{
ggplot_nlp(
  dat,
  estim,
  method_var,
  true,
  step_factors,
  pointsize = 2,
  point_dodge = 0.75,
  text_size = 4,
  top_step = NULL,
  height_betw_steps = NULL,
  height_steps = NULL,
  step_labels = NULL
)
}
\arguments{
\item{dat}{Data frame with simulation results}

\item{estim}{Character for column containing estimates}

\item{method_var}{Character for column containing method}

\item{true}{Character for column containing true data-generating 
estimand value, or a scalar e.g. 0}

\item{step_factors}{Vector of variable names to use as step factors at 
bottom of the NLP.}

\item{pointsize}{Size of points}

\item{point_dodge}{Measure of horizontal dodge}

\item{text_size}{Size of the step labels}

\item{top_step}{Position of top step (on scale of data)}

\item{height_betw_steps}{Height between NLP step (on scale of data)}

\item{height_steps}{Height of the NLP steps themselves (on scale of data)}

\item{step_labels}{Optional named vector of labels for the step functions.
Defaults to variable name and the factor labels}
}
\value{
A ggplot2 object
}
\description{
Custom ggplot-based nested loop plots
}
\examples{

sim_results <- CauseSpecCovarMI::regr_results

# Plot example code here...
}
