% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction-helpers.R
\name{haz_weib}
\alias{haz_weib}
\title{Compute Weibull hazard}
\usage{
haz_weib(alph, lam, t)
}
\arguments{
\item{alph}{Shape of Weibull distribution}

\item{lam}{Rate of Weibull distribution}

\item{t}{Postive scalar or numeric vector, for time}
}
\value{
Weibull hazard(s) at t as scalar or vector
}
\description{
Compute Weibull hazard
}
