% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_generation_helpers.R
\name{induce_missings}
\alias{induce_missings}
\title{Induce missingess.}
\usage{
induce_missings(n, .data, p, mech, eta1)
}
\arguments{
\item{n}{Sample size.}

\item{.data}{Data fram containing X and Z}

\item{p}{Proportion of missing values in X.}

\item{mech}{Missingness mechanism, one of:
"MAR_GEN", "MAR", "MNAR", "MCAR"}

\item{eta1}{Only necessary for mech != "MCAR" - degree/direction
of assocation between variable responsible for missingness
and probability of missingness}
}
\value{

}
\description{
Induce missingess.
}
