% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/illustrative-analysis-helpers.R
\name{inv_cloglog}
\alias{inv_cloglog}
\title{Inverse of complementary log-log transformation}
\usage{
inv_cloglog(x)
}
\arguments{
\item{x}{Scalar or numeric vector}
}
\description{
Inverse of complementary log-log transformation
}
