% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_generation_helpers.R
\name{logreg_missings}
\alias{logreg_missings}
\title{Calculate probability of missingness using 
logistic regression.}
\usage{
logreg_missings(p, eta1, covar)
}
\arguments{
\item{p}{Proportion of missing values in X.}

\item{eta1}{Only necessary for mech != "MCAR" - degree/direction
of assocation between variable responsible for missingness
and probability of missingness}

\item{covar}{Variable responsible for missingness, e.g. "Z"}
}
\value{

}
\description{
Calculate probability of missingness using 
logistic regression.
}
