% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstate_helpers.R
\name{make_covar_grid}
\alias{make_covar_grid}
\title{Make grid of covariate combinations.}
\usage{
make_covar_grid(dat)
}
\arguments{
\item{dat}{Dataset as generated by generate_dat()}
}
\value{
Grid of covariate combos.
}
\description{
Make grid of covariate combinations.
}
