% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/illustrative-analysis-helpers.R
\name{make_mstate_refpat}
\alias{make_mstate_refpat}
\title{Prepare reference patient to predict with probtrans}
\usage{
make_mstate_refpat(refpat, tmat, covs)
}
\arguments{
\item{refpat}{A single-row data.frane containing the covariate
values to predict}

\item{tmat}{Transition matrix}

\item{covs}{Covariates used in the cause-specific Cox models}
}
\description{
Utility function which allows to prepare one row of data.frame
to be fed into \code{mstate::msfit} for prediction.
}
\examples{
tmat <- mstate::trans.comprisk(2, c("REL", "NRM"))
covs <- c("age_allo1_decades", "mdsclass")

ref_pat <- dat_mds_synth[1, ]
ref_pat[]
make_mstate_refpat(ref_pat, tmat, covs)

}
