% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{mds_shape_rates}
\alias{mds_shape_rates}
\title{Estimated baseline hazard parameters from MDS data}
\format{
\describe{
\item{state}{Event for which we would like to generate event times, either REL
(relapse), NRM (non-relapse mortality) or EFS (event-free)}
\item{shape}{Shape of Weibull distribution}
\item{rate}{Rate of Weibull distribution}
}
}
\usage{
mds_shape_rates
}
\description{
Estimated baseline hazard parameters from MDS data
}
\keyword{datasets}
