% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MI_helpers.R
\name{mice_pool_diffm}
\alias{mice_pool_diffm}
\title{Pool different m mice}
\usage{
mice_pool_diffm(imps, m, label)
}
\arguments{
\item{m}{vector c(5, 10, 25)}

\item{label}{Label of analysis type}
}
\value{
Returns pooled estimates for biggest m and others.
}
\description{
Takes a mice() object with m imputations eg. 100
}
