% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-generation.R
\name{nelsaalen_timefixed}
\alias{nelsaalen_timefixed}
\title{Compute Nelson Aalen estimate for simulated data}
\usage{
nelsaalen_timefixed(dat, timevar, statusvar, timefix = FALSE)
}
\description{
Version of `mice::nelsonaalen` but with the
`control = survival::coxph.control(timefix = FALSE)` argument added
to the `survival::coxph` call. This is to deal with floating point errors
within the simulation procedure, where two or more simulated time points
with minimal difference between would normally be considered as a tie.
}
