% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one-simulation.R
\name{one_simulation}
\alias{one_simulation}
\title{Run one replication of one scenario}
\usage{
one_simulation(scenario, rep_num)
}
\arguments{
\item{scenario}{A row from the scenarios dataframe (link)}

\item{rep_num}{Scalar, replication numbr}
}
\description{
Run one replication of one scenario
}
