% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_generation_helpers.R
\name{pbiserial_to_pearson}
\alias{pbiserial_to_pearson}
\title{Obtain necessary pearson to generate binary variable.}
\usage{
pbiserial_to_pearson(p, r_pb)
}
\arguments{
\item{p}{Proportion of 1s in binary variable}

\item{r_pb}{Desired point-biserial correlation}
}
\value{
Pearson r to use when generating MVN data.
}
\description{
Obtain necessary pearson to generate binary variable.
}
