% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MI_helpers.R
\name{pool_diffm}
\alias{pool_diffm}
\title{Pool for different m}
\usage{
pool_diffm(mods_impdats, n_imp, analy)
}
\arguments{
\item{mods_impdats}{complete}

\item{n_imp}{Vector of m imputations to extract}

\item{analy}{String label to attach}
}
\description{
Pool for different m
}
