% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation_helpers.R
\name{pool_diffm_preds}
\alias{pool_diffm_preds}
\title{Pool for different m - with probabilities}
\usage{
pool_diffm_preds(preds_impdats, n_imp, analy)
}
\arguments{
\item{preds_impdats}{List of lalala}

\item{n_imp}{Vector of m imputations to extract}

\item{analy}{String label to attach}
}
\description{
Pool for different m - with probabilities
}
