% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/illustrative-analysis-helpers.R
\name{pool_morisot}
\alias{pool_morisot}
\title{Pooling probabilities based on complementary log-log transformation}
\usage{
pool_morisot(preds_list, by_vars)
}
\arguments{
\item{preds_list}{A list of length equal to number of imputed datasets,
containing the imputation-specific predictions. Each element should be a dataframe
containing columns "prob" (probability), "se" (standard error of probability) and 
any other variables which identify groups of predictions (to be used in by_vars)}

\item{by_vars}{Vector of variable names to pool across}
}
\description{
Method described in paper of morisot and colleagues
}
