% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/illustrative-analysis-helpers.R
\name{predict_mds_model}
\alias{predict_mds_model}
\title{Helper function to obtain predictions in the illustrative analysis}
\usage{
predict_mds_model(mod, ref_pats, tmat, horizon)
}
\arguments{
\item{mod}{Cox model based on data prepared with mstate}

\item{ref_pats}{List of reference patients as prepared by function}

\item{tmat}{Transition matrix}

\item{horizon}{Scale, time horizon of prediction}
}
\description{
(Not for use beyond this repository)
}
