% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstate_helpers.R
\name{preds_mstate}
\alias{preds_mstate}
\title{Predicting grids}
\usage{
preds_mstate(cox_long, grid_obj, times, ev1_pars, ev2_pars)
}
\arguments{
\item{cox_long}{Cox model returned by setup_mstate}

\item{times}{Prediction horizon(s) eg. c(2, 5, 10)}
}
\value{
Long cox model
}
\description{
...
}
