% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{preds_results}
\alias{preds_results}
\title{Summarised simulation results - predictions}
\format{
\describe{
\item{analy}{Method indicator, please refer to manuscript for abbreviations used}
\item{m}{Number of imputed datasets}
\item{combo-X_Z}{Covariate combination defining reference patient for which
we predict}
\item{times}{Prediction horizon (in years)}
\item{state}{State for which we estimate probability at times}
\item{true}{True data-generating cumulative incidence at a horizon, and for 
a certain reference patient}
\item{scen_num}{Number label for scenario}
\item{n}{Sample size in simulated datasets}
\item{prob}{Estimated probability}
\item{emp_se}{Empirical SE of the probabilities}
\item{bias}{Bias}
\item{rmse}{Coverage}
\item{rmse_mcse}{Coverage}
\item{bias_mcse}{Monte carlo error of bias}
\item{prop_miss}{Proportion of missing values in X}
\item{beta1}{Data-generating ffect of X on event 1}
\item{miss_mech}{Missingness mechanism}
\item{X_level}{Measurement level of X}
\item{rho}{Correlation between X and Z}
\item{eta1}{Strength of missingness mechanism for non-MCAR scenarios}
\item{haz_shape}{Shapes of baseline hazard for competing events, either "similar"
 or "different"}
}
}
\usage{
preds_results
}
\description{
Summarised simulation results - predictions
}
\keyword{datasets}
