% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MI_helpers.R
\name{quiet}
\alias{quiet}
\title{Silence function printing.}
\usage{
quiet(expr)
}
\arguments{
\item{expr}{Expression to silence

Source: http://r.789695.n4.nabble.com/Suppressing-output-e-g-from-cat-td859876.html}
}
\description{
Takes a function/expression that by default uses 
print() or cat(), and stops it. We use for MIcombine() or smcfc()
}
