% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation_helpers.R
\name{record_warning}
\alias{record_warning}
\title{Record rejection sampling failure smcfcs}
\usage{
record_warning(expr)
}
\arguments{
\item{expr}{Expression from which to record possible error}
}
\value{
List with expression and associated warning
}
\description{
Record rejection sampling failure smcfcs
}
