% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/illustrative-analysis-helpers.R
\name{reflevels_add_summary}
\alias{reflevels_add_summary}
\title{Utility to add reference factor levels to add model summary}
\usage{
reflevels_add_summary(summ, dat, form, term_col = "term")
}
\arguments{
\item{summ}{Model summary (possibly custom made)}

\item{dat}{Original dataset used to run the model}

\item{form}{Formula from the run model}

\item{term_col}{Column referencing coefficient names in summ}
}
\description{
(Not for use beyond this repository)
}
