% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{regr_results}
\alias{regr_results}
\title{Summarised simulation results - regression coefficients}
\format{
\describe{
\item{var}{Variable for which a regression coefficient is being estimated, for example
"X.1" corresponds to the cause-specific effect of X on event 1 (Relapse)}
\item{m}{Number of imputed datasets}
\item{analy}{Method indicator, please refer to manuscript for abbreviations used}
\item{true}{True data-generating value for regression coefficient}
\item{scen_num}{Number label for scenario}
\item{n}{Sample size in simulated datasets}
\item{est}{Estimated regression coefficient}
\item{se}{Estimate model standard error (SE) for regression coefficient}
\item{se_mcse}{Monte carlo standard error for model SE}
\item{emp_se}{Empirical SE}
\item{cover}{Coverage}
\item{bias}{Bias}
\item{rmse}{Coverage}
\item{rmse_mcse}{Coverage}
\item{warns}{Number of warnings across all simulations of a particular scenario, for 
a coefficient. Generally corresponds to number of smcfcs rejection sampling failures.}
\item{bias_mcse}{Monte carlo error of bias}
\item{cover_mcse}{Monte carlo error of coverage}
\item{prop_miss}{Proportion of missing values in X}
\item{beta1}{Data-generating ffect of X on event 1}
\item{miss_mech}{Missingness mechanism}
\item{X_level}{Measurement level of X}
\item{rho}{Correlation between X and Z}
\item{eta1}{Strength of missingness mechanism for non-MCAR scenarios}
\item{haz_shape}{Shapes of baseline hazard for competing events, either "similar"
 or "different"}
}
}
\usage{
regr_results
}
\description{
Summarised simulation results - regression coefficients
}
\keyword{datasets}
