% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rmse_mcse}
\alias{rmse_mcse}
\title{Approximate jackknife estimate for MCSE of RMSE}
\source{
\link{https://github.com/meghapsimatrix/simhelpers/blob/master/R/calc_absolute.R}
}
\usage{
rmse_mcse(estimates, true, K)
}
\description{
Adapted from the simhelpers CRAN package, see \link{https://cran.r-project.org/web/packages/simhelpers/vignettes/MCSE.html}.
This version takes vectors as inputs instead.
}
