% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/illustrative-analysis-helpers.R
\name{run_mds_model}
\alias{run_mds_model}
\title{Helper function to run models in the illustrative analysis}
\usage{
run_mds_model(form, tmat, dat)
}
\arguments{
\item{form}{Formula of model}

\item{tmat}{Transition matrix}

\item{dat}{Dataframe for analysis}
}
\description{
(Not for use beyond this repository)
}
