% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_generation_helpers.R
\name{rweibull_KM}
\alias{rweibull_KM}
\title{Sample from weibull in K&M parametrisation.}
\usage{
rweibull_KM(n, alph, lam)
}
\arguments{
\item{n}{sample size}

\item{alph}{Shape parameter of weibull distribution.}

\item{lam}{Rate parameter of lambda distirbution.}
}
\value{
n samples from weibull distribution.
}
\description{
Sample from weibull in K&M parametrisation.
}
