% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{scenarios}
\alias{scenarios}
\title{Scenarios table for simulation study}
\format{
\describe{
\item{n}{Sample size in simulated datasets}
\item{prop_miss}{Proportion of missing values in X}
\item{beta1}{Data-generating ffect of X on event 1}
\item{miss_mech}{Missingness mechanism}
\item{X_level}{Measurement level of X}
\item{rho}{Correlation between X and Z}
\item{eta1}{Strength of missingness mechanism for non-MCAR scenarios}
\item{haz_shape}{Shapes of baseline hazard for competing events, either "similar"
 or "different"}
\item{pilot}{Indicator of whether scenario is part of initial "pilot" scenarios
which were run to trial the simulations. Corresponds to first 15 scenarios, and 
is not relevant for the reader}
\item{scen_num}{Number label for a scenario - helps in defining the seeds}
\item{seed}{Base seed (used in \code{set.seed}) for the first replication of 
a scenario. The seed used for any other replication is then this base seed + 
replication number}
}
}
\usage{
scenarios
}
\description{
Contains all 119 scenario from simulation study as generated by
\code{analysis/simulations/make-scenarios.R}.
}
\keyword{datasets}
