% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation-helpers.R
\name{set_mi_methods}
\alias{set_mi_methods}
\title{Set methods vector for mice}
\usage{
set_mi_methods(dat, var_names_miss, imp_type = "mice", cont_method = "norm")
}
\arguments{
\item{dat}{A dataframe to be imputed}

\item{var_names_miss}{Vector of characters names for variables with missing data}

\item{imp_type}{Imputation type, either "mice" or "smcfcs"}

\item{cont_method}{Method use to impute continuous covariates, default is 
"norm" but can for example change to "pmm"}
}
\description{
Set methods vector for mice
}
