% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MI_helpers.R
\name{summarise_bayes}
\alias{summarise_bayes}
\title{Format JointAI output for sim summary}
\usage{
summarise_bayes(mod)
}
\arguments{
\item{mod}{JointAI model}
}
\value{
Formatted data.frame
}
\description{
Format JointAI output for sim summary
}
