% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation_helpers.R
\name{summarise_ref_CCA}
\alias{summarise_ref_CCA}
\title{Format CCA and ref analyses}
\usage{
summarise_ref_CCA(mod, analy)
}
\arguments{
\item{mod}{Cause-specific model}

\item{analy}{String label to attach}
}
\value{
Formatted df
}
\description{
Format CCA and ref analyses
}
