% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/illustrative-analysis-helpers.R
\name{choose_standard_refpat}
\alias{choose_standard_refpat}
\title{Choose standard reference patient to predict}
\usage{
choose_standard_refpat(
  col,
  contin_action = c("median", "mean"),
  categ_action = c("most_common", "reference")
)
}
\description{
Based on either most common or reference factor levels in sample for
categorical variable, or median/mean for continuous variables.
}
