% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/illustrative-analysis-helpers.R
\name{extract_rhs_varnames}
\alias{extract_rhs_varnames}
\title{Extract variable names from right side of model formula}
\usage{
extract_rhs_varnames(form, dat)
}
\arguments{
\item{form}{Formula object}

\item{dat}{Dataframe on which the formula is applied}
}
\value{
Character vector of variable names in right hand side of formula
}
\description{
Extract variable names from right side of model formula
}
