% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-generation.R
\name{gen_cmprsk_times}
\alias{gen_cmprsk_times}
\title{Generate competing risks times + event indictor (2 events)}
\usage{
gen_cmprsk_times(n, dat, ev1_pars, ev2_pars, rate_cens, mod_type = "latent")
}
\arguments{
\item{n}{Sample size.}

\item{dat}{Data frame containing n rows with X and Z as columns}

\item{ev1_pars}{Named list of ("a1", "h1_0", "b1", "gamm1")}

\item{ev2_pars}{Named list of ("a2", "h2_0", "b2", "gamm2")}

\item{rate_cens}{Rate of exponential distributiion for censoring.
If 0 means no censoring is applied.}

\item{mod_type}{Either "latent", weibull times generated from 
separate weibull distribution, or "total", times generated from 
sum of cause-specific hazards (using inverse transform method)}
}
\value{
Data-frame with missings induced
}
\description{
Generate competing risks times + event indictor (2 events)
}
