% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-generation.R
\name{nelsaalen_timefixed}
\alias{nelsaalen_timefixed}
\title{Compute Nelson Aalen estimate for simulated data}
\usage{
nelsaalen_timefixed(dat, timevar, statusvar, timefix = FALSE)
}
\arguments{
\item{dat}{Data frame with time and status variables}

\item{timevar}{Column name of time variable}

\item{statusvar}{Column name of (numeric) status variable}

\item{timefix}{Logical, whether to apply timefix or not. If TRUE it corresponds 
to using \code{mice::nelsonaalen}}
}
\description{
Version of \code{mice::nelsonaalen} but with the
\code{control = survival::coxph.control(timefix = FALSE)} argument added
to the \code{survival::coxph} call. This is to deal with floating point errors
within the simulation procedure, where two or more simulated time points
with minimal difference between would normally be considered as a tie.
}
