% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-generation.R
\name{rweibull_KM}
\alias{rweibull_KM}
\title{Sample from Weibull distirbution in K&M parametrisation.}
\usage{
rweibull_KM(n, alph, lam)
}
\arguments{
\item{n}{sample size}

\item{alph}{Shape parameter of weibull distribution.}

\item{lam}{Rate parameter of lambda distirbution.}
}
\value{
n samples from weibull distribution.
}
\description{
Sample from Weibull distirbution in K&M parametrisation.
}
\examples{

set.seed(4321)

times <- rweibull_KM(n = 1000, alph = 0.58, lam = 0.19)
hist(times, breaks = 30)

}
